Apply url.fileURLToPath to main.test.js
Should've been part of commit 34922071c738f2b71ecbca3beba348064fcab930.

This fixes the problem of not finding the `jsdoc` stub on Windows, but
now the tests in main.test.js are failing on Windows with:

  Serialized Error: { errno: -4094, code: 'UNKNOWN', syscall: 'spawn' }

Which is odd, because runJsdoc() calls spawn() and runJsdoc.test.js
passes just fine. Investigating...